requirements.txt
flask==3.0.3
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-cloud-vision==3.7.2
google-cloud-storage==2.17.0
requests==2.32.3
python-dotenv==1.0.1
bcrypt==4.1.3
pytest==8.2.2
gunicorn==22.0.0
